Add files via upload
The war between Russia and Ukraine is the hottest world discussion right now, since many months. Many countries are supporting Ukraine by introducing economic sanctions on Russia. There are a lot of tweets about the Ukraine and Russia war where people tend to update about the ground truths, what they feel about it, and who they are supporting. So this is a project to analyze the sentiments of people over the Ukraine and Russia. In this project, I will do Ukraine and Russia war Twitter Sentiment Analysis using Python.